requirements.txt
streamlit
groq
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
fastapi
uvicorn
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-multipart
psycopg2-binary
